DEL maxSdkVersion from WRITE_EXTERNAL permission
